EventHandler Convenience | Better connection between vars and their owners | Editor stability |
- VariableDatas and Muscariables carry arround integers connecting them to their
owners, making it easier to sustain that connection in the editor
  - The integers are connected to the owners' guids.
  - Now we don't need Muscariables or VariableDatas to carry an extra 100 bytes each
- Changed numeric types of some classes' IDs
  - Block and Command ItemIds: ints -> ushorts
  - Var Ids: ints -> bytes
- EventHandler refactored not just for rehydration, but to make it easier for subclasses
to listen and unlisten for events
- Small fixes such as removing the Undo churn when dealing with VariableDatas
- Cleanup

+++ Assets/Amanita/Scripts/Utils/IHasItemID.cs
@@ -2,6 +2,6 @@ namespace Amanita
 {
     public interface IHasItemID
     {
-        int ItemId { get; set; }
+        byte ItemId { get; set; }
     }
 }
